requirements.txt
datasets>=2.16.0
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.2
numpy>=1.24.0
tqdm>=4.66.1

langchain>=0.1.0
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
langgraph>=0.0.26
# langchain-openai>=0.0.2.post1
langchain-groq

pydantic>=2.5.3
PyYAML>=6.0.1
python-dotenv>=1.0.0             
Django>=5.0.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
PyJWT>=2.8.0